Bug Fixes and minor improvements

+++ R-Line_Courier_System/R-Line_Courier_System/frmMaintainDeliveries.cs
@@ -35,15 +35,26 @@ namespace R_Line_Courier_System
         private void btnDeleteDelivery_Click(object sender, EventArgs e)
         {
             if (dgvDelivery.Rows.Count == 2) { btnDeleteDelivery.Enabled = false; }
-            updateParcelDeliveryID(dgvDelivery.SelectedCells[0].Value.ToString());
-            con = new SqlConnection(conString);
 
-            con.Open();
-            SqlCommand cmd = new SqlCommand("DELETE FROM DELIVERIES WHERE Delivery_ID=" + dgvDelivery.SelectedCells[0].Value.ToString(), con);
-            //        try { cmd.ExecuteNonQuery(); } catch (InvalidConstraintException m) { MessageBox.Show("Delivery contains parcels"); } catch (SqlException m) { MessageBox.Show("Delivery contains parcels"); }
-            cmd.ExecuteNonQuery();
-            dataChange();
-            con.Close();
+            DialogResult res = MessageBox.Show("Are you sure you want to Delete", "Confirmation", MessageBoxButtons.OKCancel, MessageBoxIcon.Information);
+            if (res == DialogResult.OK)
+            {
+                updateParcelDeliveryID(dgvDelivery.SelectedCells[0].Value.ToString());
+                con = new SqlConnection(conString);
+
+                con.Open();
+                SqlCommand cmd = new SqlCommand("DELETE FROM DELIVERIES WHERE Delivery_ID=" + dgvDelivery.SelectedCells[0].Value.ToString(), con);
+                //        try { cmd.ExecuteNonQuery(); } catch (InvalidConstraintException m) { MessageBox.Show("Delivery contains parcels"); } catch (SqlException m) { MessageBox.Show("Delivery contains parcels"); }
+                cmd.ExecuteNonQuery();
+                dataChange();
+                con.Close();
+            }
+            if (res == DialogResult.Cancel)
+            {
+          
+            }
+
+
         }
 
         private void updateParcelDeliveryID(string deliveryID) {
